+++ dtogen/src/main/java/io/github/manoelcampos/dtogen/instantiation/RecordInstantiation.java
@@ -1,14 +1,18 @@
-package io.github.manoelcampos.dtogen;
+package io.github.manoelcampos.dtogen.instantiation;
 
+import io.github.manoelcampos.dtogen.RecordGenerator;
+
+import javax.lang.model.element.TypeElement;
 import javax.lang.model.element.VariableElement;
 import java.util.stream.Stream;
 
 /**
  * Generates the code to instantiate a record.
  * @author Manoel Campos
+ * @see ClassInstantiation#newInstance(RecordGenerator, TypeElement)
  */
 public final class RecordInstantiation extends ObjectInstantiation {
-    public RecordInstantiation(final RecordGenerator recordGen) {
+    RecordInstantiation(final RecordGenerator recordGen) {
         super(recordGen, ", ");
     }
 
@@ -19,7 +23,7 @@ public final class RecordInstantiation extends ObjectInstantiation {
 
     @Override
     protected String newObjectInternal(final String idFieldValue, final String classTypeName, final Stream<VariableElement> fieldStream) {
-        return "new %s(%s)".formatted(classTypeName, recordGen.generateFieldListInitialization(fieldStream, idFieldValue));
+        return "new %s(%s)".formatted(classTypeName, generateFieldListInitialization(typeUtil, fieldStream, idFieldValue));
     }
 
 }
